Update application path in codedeploy scripts

+++ aws/scripts/application-stop.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -xe
 
-cd ~/app/personal-blog
+cd /app/personal-blog
 
-docker compose down
+[ -f /app/personal-blog/docker-compose.yml ] && docker compose down
